Clean-up Pass.h, move viewer specific code to viewer/PassViewer.h
